+++ user-api/src/main/java/org/ecommerce/userapi/config/BeanConfiguration.java
@@ -1,6 +1,6 @@
 package org.ecommerce.userapi.config;
 
-import org.ecommerce.userapi.security.JwtProvider;
+import org.ecommerce.userapi.security.CustomAuthProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.ProviderManager;
@@ -15,7 +15,7 @@ public class BeanConfiguration {
 	}
 
 	@Bean
-	public ProviderManager providerManager(JwtProvider jwtProvider) {
-		return new ProviderManager(jwtProvider);
+	public ProviderManager providerManager(CustomAuthProvider authProvider) {
+		return new ProviderManager(authProvider);
 	}
 }
