+++ user-api/src/main/java/org/ecommerce/userapi/config/WebMvcConfig.java
@@ -2,7 +2,7 @@ package org.ecommerce.userapi.config;
 
 import java.util.List;
 
-import org.ecommerce.userapi.security.JwtUtils;
+import org.ecommerce.userapi.security.JwtProvider;
 import org.ecommerce.userapi.security.custom.CurrentUserArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
@@ -14,10 +14,10 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class WebMvcConfig implements WebMvcConfigurer {
 
-	private final JwtUtils jwtUtils;
+	private final JwtProvider jwtProvider;
 
 	@Override
 	public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-		resolvers.add(new CurrentUserArgumentResolver(jwtUtils));
+		resolvers.add(new CurrentUserArgumentResolver(jwtProvider));
 	}
 }
