+++ user-api/src/main/java/org/ecommerce/userapi/dto/UserDto.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 import org.ecommerce.userapi.entity.enumerated.Gender;
 import org.ecommerce.userapi.entity.enumerated.UserStatus;
 import org.ecommerce.userapi.exception.UserErrorMessages;
-import org.ecommerce.userapi.security.JwtUtils;
+import org.ecommerce.userapi.security.JwtProvider;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -81,7 +81,7 @@ public class UserDto {
 			String accessToken
 		) {
 			public static Login of(final UserDto userDto) {
-				return new Login(JwtUtils.prefix(userDto.accessToken));
+				return new Login(JwtProvider.prefix(userDto.accessToken));
 			}
 		}
 	}
