+++ user-api/src/main/java/org/ecommerce/userapi/entity/Users.java
@@ -17,7 +17,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -26,7 +25,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
 public class Users {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
