+++ user-api/src/main/java/org/ecommerce/userapi/repository/SellerAccountRepository.java
@@ -3,5 +3,5 @@ package org.ecommerce.userapi.repository;
 import org.ecommerce.userapi.entity.SellerAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SellerAccountRepository extends JpaRepository<SellerAccount, Integer> {
+public interface SellerAccountRepository extends JpaRepository<SellerAccount, Integer>, SellerCustomRepository {
 }
