+++ user-api/src/main/java/org/ecommerce/userapi/security/JwtProvider.java
@@ -1,51 +1,177 @@
 package org.ecommerce.userapi.security;
 
-import java.util.HashSet;
+import java.util.Date;
+import java.util.List;
 import java.util.Set;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+
+import javax.crypto.SecretKey;
 
 import org.ecommerce.common.error.CustomException;
 import org.ecommerce.userapi.exception.UserErrorCode;
-import org.springframework.security.authentication.AuthenticationProvider;
+import org.ecommerce.userapi.utils.RedisProvider;
+import org.springframework.http.HttpHeaders;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Component;
 
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.MalformedJwtException;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.UnsupportedJwtException;
+import io.jsonwebtoken.security.SignatureException;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class JwtProvider implements AuthenticationProvider {
+public class JwtProvider {
+	private static final String ACCESS_TOKEN_HEADER = HttpHeaders.AUTHORIZATION;
+	private static final String PREFIX = "Bearer ";
+	private static final long ONE_HOUR = 3_600;
+	private static final long TWO_WEEKS = ONE_HOUR * 24 * 14;
 
-	private final JwtUtils jwtUtils;
+	private final RedisProvider redisProvider;
 
-	@Override
-	public Authentication authenticate(Authentication authentication) throws AuthenticationException {
+	private final SecretKey secretKey;
 
-		String email = (String)authentication.getPrincipal();
-		String bearerToken = (String)authentication.getCredentials();
-
-		String parseEmail = jwtUtils.getEmail(bearerToken);
-		String parseRole = jwtUtils.getRoll(bearerToken);
-		Integer parseId = jwtUtils.getId(bearerToken);
-
-		AuthDetails authDetails = new AuthDetails(parseId, parseEmail, parseRole);
-
-		SimpleGrantedAuthority parsedGrant = new SimpleGrantedAuthority(parseRole);
-
-		Set<SimpleGrantedAuthority> authorities = new HashSet<>();
-		authorities.add(parsedGrant);
-
-		if (!authDetails.getEmail().equals(email) ||
-			!authentication.getAuthorities().contains(parsedGrant)) {
-			throw new CustomException(UserErrorCode.AUTHENTICATION_FAILED);
-		}
-		return new UsernamePasswordAuthenticationToken(authDetails, bearerToken, authorities);
+	public static String prefix(String jwt) {
+		if (jwt.startsWith(PREFIX))
+			return jwt;
+		return PREFIX + jwt;
 	}
 
-	@Override
-	public boolean supports(Class<?> authentication) {
-		return authentication.equals(UsernamePasswordAuthenticationToken.class);
+	public String createSellerTokens(Integer sellerId, String email, Set<String> authorization,
+		HttpServletResponse response) {
+
+		final String accessToken = createToken(sellerId, email,
+			ONE_HOUR, secretKey, authorization);
+		final String refreshToken = createToken(sellerId, email,
+			TWO_WEEKS, secretKey, authorization);
+
+		String accessTokenKey = getAccessTokenKey(sellerId, getRoll(accessToken));
+		redisProvider.setData(accessTokenKey, accessToken, ONE_HOUR, TimeUnit.SECONDS);
+
+		String refreshTokenKey = getRefreshTokenKey(sellerId, getRoll(refreshToken));
+		redisProvider.setData(refreshTokenKey, refreshToken, TWO_WEEKS, TimeUnit.SECONDS);
+
+		response.addCookie(createCookie(refreshToken));
+
+		return accessToken;
+	}
+
+	public String createUserTokens(Integer userId, String email, Set<String> authorization,
+		HttpServletResponse response) {
+
+		final String accessToken = createToken(userId, email,
+			ONE_HOUR, secretKey, authorization);
+		final String refreshToken = createToken(userId, email,
+			TWO_WEEKS, secretKey, authorization);
+
+		String accessTokenKey = getAccessTokenKey(userId, getRoll(accessToken));
+		redisProvider.setData(accessTokenKey, accessToken, ONE_HOUR, TimeUnit.SECONDS);
+
+		String refreshTokenKey = getRefreshTokenKey(userId, getRoll(refreshToken));
+		redisProvider.setData(refreshTokenKey, refreshToken, TWO_WEEKS, TimeUnit.SECONDS);
+		
+		response.addCookie(createCookie(refreshToken));
+
+		return accessToken;
+	}
+
+	private String createToken(Integer id, String email, long lifeCycle, SecretKey secretKey,
+		Set<String> authorization) {
+		Date now = new Date();
+		return Jwts.builder()
+			.claim("id", id)
+			.claim("email", email)
+			.claim("authorization", authorization)
+			.setIssuedAt(now)
+			.setExpiration(new Date(now.getTime() + lifeCycle * 1000))
+			.signWith(SignatureAlgorithm.HS256, secretKey)
+			.compact();
+	}
+
+	public Claims parseClaims(String jwt) throws CustomException {
+		try {
+			return Jwts.parserBuilder().setSigningKey(secretKey).build().parseClaimsJws(jwt).getBody();
+		} catch (ExpiredJwtException | UnsupportedJwtException | MalformedJwtException | SignatureException |
+				 IllegalArgumentException e) {
+			throw e;
+		}
+	}
+
+	public UsernamePasswordAuthenticationToken parseAuthentication(String bearerToken) {
+		Claims claims = parseClaims(cutPreFix(bearerToken));
+		String email = claims.get("email", String.class);
+		List<?> authorities = claims.get("authorization", List.class);
+
+		if (authorities == null)
+			throw new CustomException(UserErrorCode.INVALID_SIGNATURE_JWT);
+
+		Set<SimpleGrantedAuthority> grantedAuthorities = authorities.stream()
+			.map(String::valueOf)
+			.map(SimpleGrantedAuthority::new)
+			.collect(Collectors.toSet());
+
+		return new UsernamePasswordAuthenticationToken(email, bearerToken, grantedAuthorities);
+	}
+
+	public void removeTokens(String accessTokenKey, String refreshTokenKey) {
+		redisProvider.deleteData(accessTokenKey);
+		redisProvider.deleteData(refreshTokenKey);
+	}
+
+	public String getRoll(String bearerToken) {
+		return parseClaims(cutPreFix(bearerToken)).get("authorization", List.class).get(0).toString();
+	}
+
+	public String getEmail(String bearerToken) {
+		return parseClaims(cutPreFix(bearerToken)).get("email", String.class);
+	}
+
+	public Integer getId(String bearerToken) {
+		return parseClaims(cutPreFix(bearerToken)).get("id", Integer.class);
+	}
+
+	public String getRefreshTokenKey(Integer id, String auth) {
+		return "refresh_token" + id + auth;
+	}
+
+	public String getAccessTokenKey(Integer id, String auth) {
+		return "access_token" + id + auth;
+	}
+
+	public String cutPreFix(String bearerToken) {
+		if (!bearerToken.startsWith(PREFIX))
+			return bearerToken;
+		return bearerToken.substring(PREFIX.length());
+	}
+
+	public String resolveToken(HttpServletRequest request) {
+		return request.getHeader(ACCESS_TOKEN_HEADER);
+	}
+
+	public Cookie createRefreshTokenCookie(String refreshToken) {
+		Cookie cookie = new Cookie("refreshToken", refreshToken);
+		cookie.setHttpOnly(true);
+		cookie.setSecure(true);
+		cookie.setPath("/");
+		cookie.setMaxAge((int)TWO_WEEKS);
+		return cookie;
+	}
+
+	private Cookie createCookie(String refreshToken) {
+		Cookie cookie = new Cookie("refreshToken", refreshToken);
+		cookie.setHttpOnly(true);
+		cookie.setSecure(true);
+		cookie.setPath("/");
+		cookie.setMaxAge((int)TWO_WEEKS);
+		return cookie;
 	}
 }
