+++ user-api/src/main/java/org/ecommerce/userapi/security/custom/CurrentUserArgumentResolver.java
@@ -3,7 +3,7 @@ package org.ecommerce.userapi.security.custom;
 import java.util.Objects;
 
 import org.ecommerce.userapi.security.AuthDetails;
-import org.ecommerce.userapi.security.JwtUtils;
+import org.ecommerce.userapi.security.JwtProvider;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -18,7 +18,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class CurrentUserArgumentResolver implements HandlerMethodArgumentResolver {
 
-	private final JwtUtils jwtUtils;
+	private final JwtProvider jwtProvider;
 
 	@Override
 	public boolean supportsParameter(MethodParameter parameter) {
@@ -28,12 +28,12 @@ public class CurrentUserArgumentResolver implements HandlerMethodArgumentResolve
 	@Override
 	public AuthDetails resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
 		NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
-		String bearerToken = jwtUtils.resolveToken(
+		String bearerToken = jwtProvider.resolveToken(
 			Objects.requireNonNull(webRequest.getNativeRequest(HttpServletRequest.class)));
 		if (bearerToken != null) {
-			Integer userId = jwtUtils.getId(bearerToken);
-			String email = jwtUtils.getEmail(bearerToken);
-			String role = jwtUtils.getRoll(bearerToken);
+			Integer userId = jwtProvider.getId(bearerToken);
+			String email = jwtProvider.getEmail(bearerToken);
+			String role = jwtProvider.getRoll(bearerToken);
 			return new AuthDetails(userId, email, role);
 		}
 		return new AuthDetails(null, null, null);
