+++ user-api/src/main/java/org/ecommerce/userapi/utils/RedisProvider.java
@@ -9,22 +9,23 @@ import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class RedisUtils {
+public class RedisProvider {
 
 	private final RedisTemplate<String, Object> redisTemplate;
 
-
-	public void setData(String key, String value,Long expiredTime, TimeUnit timeUnit){
+	public void setData(String key, String value, Long expiredTime, TimeUnit timeUnit) {
 		redisTemplate.opsForValue().set(key, value, expiredTime, timeUnit);
 	}
-	public String getData(String key){
-		return (String) redisTemplate.opsForValue().get(key);
+
+	public String getData(String key) {
+		return (String)redisTemplate.opsForValue().get(key);
 	}
 
 	public boolean hasKey(String key) {
 		return Boolean.TRUE.equals(redisTemplate.hasKey(key));
 	}
-	public void deleteData(String key){
+
+	public void deleteData(String key) {
 		redisTemplate.delete(key);
 	}
 }
