+++ user-api/src/test/java/org/ecommerce/userapi/service/SellerServiceTest.java
@@ -17,7 +17,7 @@ import org.ecommerce.userapi.external.service.SellerService;
 import org.ecommerce.userapi.repository.SellerAccountRepository;
 import org.ecommerce.userapi.repository.SellerRepository;
 import org.ecommerce.userapi.security.AuthDetails;
-import org.ecommerce.userapi.security.JwtUtils;
+import org.ecommerce.userapi.security.JwtProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -44,7 +44,7 @@ class SellerServiceTest {
 	private SellerAccountRepository sellerAccountRepository;
 
 	@Mock
-	private JwtUtils jwtUtils;
+	private JwtProvider jwtProvider;
 
 	@BeforeEach
 	public void 기초_셋팅() {
@@ -173,7 +173,7 @@ class SellerServiceTest {
 			Seller seller = Seller.ofRegister(email, "John Doe", password, "어쩌구 저쩌구", "01012345678");
 			when(sellerRepository.findByEmail(email)).thenReturn(Optional.of(seller));
 			when(bCryptPasswordEncoder.matches(password, seller.getPassword())).thenReturn(true);
-			when(jwtUtils.createSellerTokens(any(), any(), any(), any())).thenReturn("Bearer fake_access_token");
+			when(jwtProvider.createSellerTokens(any(), any(), any(), any())).thenReturn("Bearer fake_access_token");
 
 			// when
 			SellerDto expectedResponse = sellerService.loginRequest(loginRequest, response);
